refactoring (rename project microcommerce to microcommereapi)
