--- microcommerce/src/main/java/com/ecommerce/microcommerce/dao/ProductDao.java
@@ -1,14 +1,0 @@
-package com.ecommerce.microcommerce.dao;
-
-import java.util.List;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.ecommerce.microcommerce.model.Product;
-
-@Repository
-public interface ProductDao extends JpaRepository<Product, Integer> {
-
-	List<Product> findAllByOrderByNameAsc();
-}
